- Caching now transparently forwards exceptions and keeps working when serializers throw exceptions - Improved database-layer, removed 2 bytes for cid in references - General performance improvements - Model-data not stored transactionally anymore. Store model (which is transactional) is guarding access to models - Improved updating plugins - Updated to dump methods of new query objects to avoid hash collisions for different queries

+++ BimServer/src/org/bimserver/database/queries/QueryBoundingBoxStackFrame.java
@@ -97,12 +97,12 @@ public class QueryBoundingBoxStackFrame extends DatabaseReadingStackFrame implem
 				HashMapVirtualObject geometryInfo = getByOid(geometryInfoId);
 				HashMapWrappedVirtualObject minBounds = (HashMapWrappedVirtualObject) geometryInfo.get("minBounds");
 				HashMapWrappedVirtualObject maxBounds = (HashMapWrappedVirtualObject) geometryInfo.get("maxBounds");
-				float minX = (float) minBounds.eGet("x");
-				float minY = (float) minBounds.eGet("y");
-				float minZ = (float) minBounds.eGet("z");
-				float maxX = (float) maxBounds.eGet("x");
-				float maxY = (float) maxBounds.eGet("y");
-				float maxZ = (float) maxBounds.eGet("z");
+				double minX = (double) minBounds.eGet("x");
+				double minY = (double) minBounds.eGet("y");
+				double minZ = (double) minBounds.eGet("z");
+				double maxX = (double) maxBounds.eGet("x");
+				double maxY = (double) maxBounds.eGet("y");
+				double maxZ = (double) maxBounds.eGet("z");
 				if (
 						minX > inBoundingBox.getX() &&
 						minY > inBoundingBox.getY() &&
@@ -110,7 +110,6 @@ public class QueryBoundingBoxStackFrame extends DatabaseReadingStackFrame implem
 						maxX < inBoundingBox.getX() + inBoundingBox.getWidth() &&
 						maxY < inBoundingBox.getY() + inBoundingBox.getHeight() &&
 						maxZ < inBoundingBox.getZ() + inBoundingBox.getDepth()) {
-					
 				} else {
 					currentObject = null;
 				}
